Update now splits the data untill there are zero duplicates left.
Removed the logger library and removed all logger calls in the code.
Removed all print lines from the code aswell.
Reset the way the export is printed as a result, instead of the testing way where it looks like the starting data.
Added some comment to the code to make the code more understandable.
Extra check to make sure no links from records are combined if the links don't complete a parent link code.
Added update of the codesToIds and parentRelations after combining Link codes. This to make sure the data is handled correctly.
Making sure the codes to remove are really needed to remove, otherwise clears the list so no Link codes are removed on accident.
If the result values contain a value of 0, they are added like they are and added to a seperate list to later on check whether the starting value of the record is zero.
Multiple extra checks for the Link codes are added throughout the code.
Code added to split the Links of records when there are more than two Link codes present. This both for when the value is not 0 and if it is 0. When it is 0, the squarekilometres are used to calculate.