Cleaned up the code
Removed unnecesary/unused code.
Updated and/or added commentary to each method in the file.
Changed the way the records with null house numbers are added to the map of records and the way these are handled further on in the code. This to make sure the records which start with 0 homes are saved as 0 homes, and otherwise saved as N/A.
Removed duplicate code and made methods of these parts of duplicate code.