Improved record splitting
Wrote an extra check to make sure the codes collected really combine the codes that are in the links of a record. If this is not the case, part of the code is skipped.
Created extra ifs to check whether a record does or doesn't contain null as a value for the number of homes.
Check added to see whether the size of a recordlist equals to 1.
Added an if to check whether the number of homes not equals null for further handling when using square kilometres.
Created a method for some duplicate code.
Rewrote the way the village is added to dorpcollected, to make sure the links starting with GE_ are added aswell.
Added a toString method to villageComplex for debugging purposes.
Added a check to see whether a link code contains "HO" before setting the parent relations.